Arreglado, quitar guardados tras la modularizacion de los eventos en PointSummary
