+++ webapp/src/pages/saved/SavedPointsPage.tsx
@@ -5,10 +5,15 @@ import "../../public/css/pages/saved/SavedPointsPage.scss";
 import { useEffect } from "react";
 import { useAllSavedPointsStore } from "../../store/point.store";
 import PointSummaryWithMap from "../../components/points/PointSummaryWithMap";
+import { Point } from "../../shared/shareddtypes";
+import { unsavePoint } from "../../api/save.point.api";
+import { useNavigate } from "react-router-dom";
+import { HOME_PATH } from "../../routes";
 
 function SavedPointsPage() {
   const { session } = useSession();
   const { setSavedPoints, savedPoints } = useAllSavedPointsStore();
+  const navigate = useNavigate();
 
   const loadAllSavedPointsByCurrentUser = async () => {
     const allSavedPoints = await findAllSavePoints(session.info.webId as string);
@@ -18,6 +23,25 @@ function SavedPointsPage() {
   useEffect(() => {
     loadAllSavedPointsByCurrentUser();
   }, []);
+
+  const handleDeleteSavedPoint = (
+    e: React.MouseEvent<HTMLElement>,
+    pointInfo : Point,
+    ) => {
+      e.preventDefault();
+      let idPoint = '';
+      let webId = '';
+      if(pointInfo?._id){
+        idPoint = pointInfo?._id;
+      }
+      if(session.info.webId){
+        webId = session.info.webId;
+      }
+
+      unsavePoint(idPoint,webId);
+
+      navigate(HOME_PATH)
+    }
 
   return (
     <AccountLayout hasBanner={false}>
@@ -34,6 +58,7 @@ function SavedPointsPage() {
                 lng={point.location?.coords.lng || 0}
                 hasMap={true}
                 pointInfo={point}
+                onClick={(e) => handleDeleteSavedPoint(e ,point)}
               />
             ))
           ) : (
